DynamicObject refactored
