+++ A2v10.ProcS.Tests/CallApiTest.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright © 2020 Alex Kukhtin, Artur Moshkola. All rights reserved.
 
+using A2v10.ProcS.Infrastructure;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
