+++ A2v10.ProcS.Tests/ScriptActionTest.cs
@@ -2,7 +2,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Dynamic;
 using System.IO;
 using System.Threading.Tasks;
 using A2v10.ProcS.Infrastructure;
