+++ A2v10.ProcS/Activities/ParallelActivity.cs
@@ -2,7 +2,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Dynamic;
 using System.Linq;
 using A2v10.ProcS.Infrastructure;
