+++ A2v10.ProcS/Activities/StartProcessActivity.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright © 2020 Alex Kukhtin, Artur Moshkola. All rights reserved.
 
 using System;
-using System.Dynamic;
-using System.Threading.Tasks;
 using A2v10.ProcS.Infrastructure;
 
 namespace A2v10.ProcS
@@ -20,7 +18,7 @@ namespace A2v10.ProcS
 			if (context.IsContinue)
 				return ActivityExecutionResult.Complete;
 
-			var prms = context.EvaluateScript<ExpandoObject>(ParameterExpression);
+			var prms = context.EvaluateScript<System.Dynamic.ExpandoObject>(ParameterExpression);
 			var startMessage = new StartProcessMessage(context.Instance.Id)
 			{
 				ProcessId = Process,
