+++ A2v10.ProcS/StateMachine.cs
@@ -2,7 +2,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Dynamic;
 using System.Linq;
 using System.Threading.Tasks;
 using A2v10.ProcS.Infrastructure;
