+++ A2v10.ProcS/WorkflowEngine.cs
@@ -2,7 +2,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Dynamic;
 using System.Text;
 using System.Threading.Tasks;
 using A2v10.ProcS.Infrastructure;
